requirements.txt
anthropic>=0.40.0
pytrends>=4.9.2
newsapi-python>=0.2.7
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.6.0
mistune>=3.0.2
streamlit>=1.32.0
apscheduler>=3.10.4
